tests/wpt/tests/html/canvas/element/fill-and-stroke-styles/2d.gradient.conic.positive.rotation.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.gradient.conic.positive.rotation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.gradient.conic.positive.rotation</h1>
<p class="desc">Conic gradient with positive rotation</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("Conic gradient with positive rotation");
_addTest(function(canvas, ctx) {

  const g = ctx.createConicGradient(3*Math.PI/2, 50, 25);
  // It's red in the upper right region and green on the lower left region
  g.addColorStop(0, "#f00");
  g.addColorStop(0.25, "#0f0");
  g.addColorStop(0.50, "#0f0");
  g.addColorStop(0.75, "#f00");
  ctx.fillStyle = g;
  ctx.fillRect(0, 0, 100, 50);
  _assertPixelApprox(canvas, 25,15, 255,0,0,255, 3);
  _assertPixelApprox(canvas, 75,40, 0,255,0,255, 3);

});
</script>
