LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/fill-and-stroke-styles/2d.gradient.conic.positive.rotation-expected.txt

2d.gradient.conic.positive.rotation
Conic gradient with positive rotation
Actual output:
Expected output:


FAIL Conic gradient with positive rotation assert_equals: Red channel of the pixel at (25, 15) expected 255 but got 195
